airbyte-integrations/connectors/source-victoria-metrics/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-victoria-metrics"
description = "Source implementation for victoria-metrics."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/victoria-metrics"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_victoria_metrics" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.9,<3.12"
airbyte-cdk = "^0"

[tool.poetry.scripts]
source-victoria-metrics = "source_victoria_metrics.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
mypy = "1.10.0"
ruff = "0.4.4"

[tool.ruff]
# https://docs.astral.sh/ruff/
line-length = 120

[tool.ruff.lint]
select = [
    # https://docs.astral.sh/ruff/rules/
    "F",     # pyflakes
    "E",     # pycodestyle (errors)
    "W",     # pycodestyle (warnings) eg. trailing/leading whitespaces
    "Q",     # quotes
    "I001",  # unsorted imports (isort)
    "TID",   # tidy-imports
]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
# https://docs.astral.sh/ruff/rules/#isort-i

[tool.autopep8]
# docs: https://github.com/hhatto/autopep8
max_line_length = 120
aggressive = 2

[tool.mypy]
# docs: https://mypy.readthedocs.io/en/stable/config_file.html
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
warn_unused_ignores = true
ignore_errors = false
follow_imports = "silent"
ignore_missing_imports = true
strict = false

# ignore mypy errors in tests
[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true